requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# LangChain & AI
langchain==0.1.0
langchain-openai==0.0.2
openai==1.7.0

# YouTube & Web Scraping
yt-dlp==2024.1.7
requests==2.31.0
beautifulsoup4==4.12.2

# Document Processing
pytesseract==0.3.10
pdf2image==1.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Audio/Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0

# Utilities
nest-asyncio==1.5.8
Pillow==10.2.0
lxml==5.1.0

# Additional dependencies that may be needed
numpy==1.26.3
tiktoken==0.5.2